Dont use yin suicide ability without reason.
That is, dont use it if it is the last ship and all enemy ships have
sustain. It could be argued that it is still worth it if the fight is
hopeless. But the most realistic situation where this happens is like a
destroyed vs a dreadnought. So I think it is an improvement.